Create helper to specs with fillFormFields method and add specs to validateStep method
